static cast std::size_t to int within size() method

+++ Algorithms-Part1/KDTrees/include/PointSet.h
@@ -37,7 +37,7 @@ public:
      * 
      * \returns Number of points in the set.
      */
-    int size() const { return m_pset.size(); }
+    int size() const { return static_cast<int>(m_pset.size()); }
 
     /**
      * Inserts point into the set. Must be unique, no duplicate points allowed.
